\name{GeoIP_num_countries}

\alias{GeoIP_num_countries}

\title{An interface to the native routine GeoIP_num_countries}

\description{This function allows one to invoke the native routine GeoIP_num_countries from R, passing R values as arguments to the routine.}

\arguments{
\item{.cif}{the call interface object describing the signature of the native routine}
\item{\dots}{additional parameters that are passed to \code{\link[Rffi]{callCIF}}}}

\value{the native routine returns an object of class numeric. }

\keyword{programming}

\keyword{interface}
