\name{GeoIPCharsetValues}

\alias{GeoIPCharsetValues}

\alias{GEOIP_CHARSET_ISO_8859_1}

\alias{GEOIP_CHARSET_UTF8}

\alias{GeoIPCharsetValues}

\alias{coerce,numeric,GeoIPCharset-method}

\alias{coerce,integer,GeoIPCharset-method}

\alias{coerce,character,GeoIPCharset-method}

\title{Enumeration values and class for GeoIPCharset}

\description{We represent the C-level enumerations for the type GeoIPCharset with a collection of R variables and an R class that allows us to coerce from numbers and strings to the enumeration type.}
