\name{GeoIPDBTypesValues}

\alias{GeoIPDBTypesValues}

\alias{GEOIP_COUNTRY_EDITION}

\alias{GEOIP_REGION_EDITION_REV0}

\alias{GEOIP_CITY_EDITION_REV0}

\alias{GEOIP_ORG_EDITION}

\alias{GEOIP_ISP_EDITION}

\alias{GEOIP_CITY_EDITION_REV1}

\alias{GEOIP_REGION_EDITION_REV1}

\alias{GEOIP_PROXY_EDITION}

\alias{GEOIP_ASNUM_EDITION}

\alias{GEOIP_NETSPEED_EDITION}

\alias{GEOIP_DOMAIN_EDITION}

\alias{GEOIP_COUNTRY_EDITION_V6}

\alias{GeoIPDBTypesValues}

\alias{coerce,numeric,GeoIPDBTypes-method}

\alias{coerce,integer,GeoIPDBTypes-method}

\alias{coerce,character,GeoIPDBTypes-method}

\title{Enumeration values and class for GeoIPDBTypes}

\description{We represent the C-level enumerations for the type GeoIPDBTypes with a collection of R variables and an R class that allows us to coerce from numbers and strings to the enumeration type.}
