\name{GeoIPNetspeedValuesValues}

\alias{GeoIPNetspeedValuesValues}

\alias{GEOIP_UNKNOWN_SPEED}

\alias{GEOIP_DIALUP_SPEED}

\alias{GEOIP_CABLEDSL_SPEED}

\alias{GEOIP_CORPORATE_SPEED}

\alias{GeoIPNetspeedValuesValues}

\alias{coerce,numeric,GeoIPNetspeedValues-method}

\alias{coerce,integer,GeoIPNetspeedValues-method}

\alias{coerce,character,GeoIPNetspeedValues-method}

\title{Enumeration values and class for GeoIPNetspeedValues}

\description{We represent the C-level enumerations for the type GeoIPNetspeedValues with a collection of R variables and an R class that allows us to coerce from numbers and strings to the enumeration type.}
