\name{GeoIPOptionsValues}

\alias{GeoIPOptionsValues}

\alias{GEOIP_STANDARD}

\alias{GEOIP_MEMORY_CACHE}

\alias{GEOIP_CHECK_CACHE}

\alias{GEOIP_INDEX_CACHE}

\alias{GEOIP_MMAP_CACHE}

\alias{GeoIPOptionsValues}

\alias{coerce,numeric,GeoIPOptions-method}

\alias{coerce,integer,GeoIPOptions-method}

\alias{coerce,character,GeoIPOptions-method}

\title{Enumeration values and class for GeoIPOptions}

\description{We represent the C-level enumerations for the type GeoIPOptions with a collection of R variables and an R class that allows us to coerce from numbers and strings to the enumeration type.}
