\name{GeoIPProxyTypes}

\alias{GeoIPProxyTypes-class}

\alias{GeoIPProxyTypesPtr-class}

\alias{GeoIPProxyTypes}

\alias{GeoIPProxyTypesPtr}

\alias{$,GeoIPProxyTypesPtr-method}

\alias{coerce,GeoIPProxyTypesPtr,GeoIPProxyTypes-method}

\alias{coerce,GeoIPProxyTypesPtr,GeoIPProxyTypesPtr-method}

\alias{coerce,GeoIPProxyTypesPtr,externalptr-method}

\alias{coerce,externalptr,GeoIPProxyTypesPtr-method}

\alias{GeoIPProxyTypes,ANY-method}

\alias{GeoIPProxyTypes,externalptr-method}

\alias{GeoIPProxyTypes,GeoIPProxyTypesPtr-method}

\alias{GeoIPProxyTypesPtr,missing-method}

\alias{GeoIPProxyTypesPtr,externalptr-method}

\alias{names,GeoIPProxyTypesPtr-method}

\title{R classes to represent native data type GeoIPProxyTypes as R object or reference to native type}

\description{These classes and methods provide an R user with access to the native data type GeoIPProxyTypes. We define an R version of this data structure with R-level fields, and also an R class that can reference an instance of the native data type. For this reference, we define methods that make the object act as if it were a list in R so that one can access fields via the usual subsetting operators. One can also coerce from one representation to an nother and create new objects via the GeoIPProxyTypes constructor function.}

\value{The names methods returns a character vector. The constructor functions create objects of the class with the same name as the constructor. The $ operator returns an object as the same class as the field, of course.}

\keyword{programming}

\keyword{interface}
