\name{GeoIPProxyTypesValues}

\alias{GeoIPProxyTypesValues}

\alias{GEOIP_ANON_PROXY}

\alias{GEOIP_HTTP_X_FORWARDED_FOR_PROXY}

\alias{GEOIP_HTTP_CLIENT_IP_PROXY}

\alias{GeoIPProxyTypesValues}

\alias{coerce,numeric,GeoIPProxyTypes-method}

\alias{coerce,integer,GeoIPProxyTypes-method}

\alias{coerce,character,GeoIPProxyTypes-method}

\title{Enumeration values and class for GeoIPProxyTypes}

\description{We represent the C-level enumerations for the type GeoIPProxyTypes with a collection of R variables and an R class that allows us to coerce from numbers and strings to the enumeration type.}
