\name{GeoIPRegionTag}

\alias{GeoIPRegionTag-class}

\alias{GeoIPRegionTagPtr-class}

\alias{GeoIPRegionTag}

\alias{GeoIPRegionTagPtr}

\alias{$,GeoIPRegionTagPtr-method}

\alias{coerce,GeoIPRegionTagPtr,GeoIPRegionTag-method}

\alias{coerce,GeoIPRegionTagPtr,GeoIPRegionTagPtr-method}

\alias{coerce,GeoIPRegionTagPtr,externalptr-method}

\alias{coerce,externalptr,GeoIPRegionTagPtr-method}

\alias{GeoIPRegionTag,ANY-method}

\alias{GeoIPRegionTag,externalptr-method}

\alias{GeoIPRegionTag,GeoIPRegionTagPtr-method}

\alias{GeoIPRegionTagPtr,missing-method}

\alias{GeoIPRegionTagPtr,externalptr-method}

\alias{names,GeoIPRegionTagPtr-method}

\title{R classes to represent native data type GeoIPRegionTag as R object or reference to native type}

\description{These classes and methods provide an R user with access to the native data type GeoIPRegionTag. We define an R version of this data structure with R-level fields, and also an R class that can reference an instance of the native data type. For this reference, we define methods that make the object act as if it were a list in R so that one can access fields via the usual subsetting operators. One can also coerce from one representation to an nother and create new objects via the GeoIPRegionTag constructor function.}

\value{The names methods returns a character vector. The constructor functions create objects of the class with the same name as the constructor. The $ operator returns an object as the same class as the field, of course.}

\keyword{programming}

\keyword{interface}
