\name{idtype_tValues}

\alias{idtype_tValues}

\alias{P_ALL}

\alias{P_PID}

\alias{P_PGID}

\alias{idtype_tValues}

\alias{coerce,numeric,idtype_t-method}

\alias{coerce,integer,idtype_t-method}

\alias{coerce,character,idtype_t-method}

\title{Enumeration values and class for idtype_t}

\description{We represent the C-level enumerations for the type idtype_t with a collection of R variables and an R class that allows us to coerce from numbers and strings to the enumeration type.}
